requirements_rag.txt
# RAG 의존성
chromadb>=0.5.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
langchain-core>=0.3.0
